+++ packages/preset/src/_shortcuts/index.ts
@@ -20,6 +20,7 @@ import { checkbox } from './checkbox'
 import { slider } from './slider'
 import { progress } from './progress'
 import { skeleton } from './skeleton'
+import { tabs } from './tabs'
 import { select } from './select'
 import { separator } from './separator'
 import { card } from './card'
@@ -46,6 +47,7 @@ export const shortcuts = [
   ...slider,
   ...progress,
   ...skeleton,
+  ...tabs,
   ...select,
   ...separator,
   ...card,